+++ example/main.dart
@@ -7,55 +7,46 @@ import 'package:teledart/model.dart';
 
 Future<void> main() async {
   final envVars = Platform.environment;
-
-  var telegram = Telegram(envVars['BOT_TOKEN']!);
-  var event = Event((await telegram.getMe()).username!);
+  final username = (await Telegram(envVars['BOT_TOKEN']!).getMe()).username;
 
   // TeleDart uses longpoll by default if no update fetcher is specified.
-  var teledart = TeleDart(telegram, event);
+  var teledart = TeleDart(envVars['BOT_TOKEN']!, Event(username!));
 
   // In case you decided to use webhook.
   // var webhook = await Webhook.createHttpsWebhok(
-  //     telegram,
+  //     Telegram(envVars['BOT_TOKEN']!),
   //     envVars['HOST_URL']!,
   //     envVars['BOT_TOKEN']!,
   //     io.File(envVars['CERT_PATH']!),
   //     io.File(envVars['KEY_PATH']!),
   //     port: int.parse(envVars['BOT_PORT']!));
-  // var teledart = TeleDart(telegram, event, fetcher: webhook);
+  // var teledart = TeleDart(envVars['BOT_TOKEN']!, Event(username!), fetcher: webhook);
 
   teledart.start();
 
   // You can listen to messages like this
   teledart.onMessage(entityType: 'bot_command', keyword: 'start').listen(
-      (message) =>
-          teledart.telegram.sendMessage(message.chat.id, 'Hello TeleDart!'));
+      (message) => teledart.sendMessage(message.chat.id, 'Hello TeleDart!'));
 
-  // Or using short cuts
-  teledart
-      .onCommand('short')
-      .listen(((message) => teledart.replyMessage(message, 'This works too!')));
+  // Sick of boilerplates? Reply messages like below, nice and tidy
+  // Short hands also available for answer query methods
+  teledart.onCommand('short')
+    .listen((message) => message.reply('This works too!'));
 
   // You can also utilise regular expressions
-  teledart.onCommand(RegExp('hello', caseSensitive: false)).listen(
-      (message) => teledart.telegram.sendMessage(message.chat.id, 'hi!'));
+  teledart
+      .onCommand(RegExp('hello', caseSensitive: false))
+      .listen((message) => message.reply('hi!'));
 
   // You can even filter streams with stream processing methods
   // See: https://www.dartlang.org/tutorials/language/streams#methods-that-modify-a-stream
   teledart
       .onMessage(keyword: 'dart')
       .where((message) => message.text?.contains('telegram') ?? false)
-      .listen((message) => teledart.replyPhoto(
-          message,
+      .listen((message) => message.replyPhoto(
           //  io.File('example/dash_paper_plane.png'),
           'https://raw.githubusercontent.com/DinoLeung/TeleDart/master/example/dash_paper_plane.png',
           caption: 'This is how Dash found the paper plane'));
-
-  // Sick of boilerplates? Reply messages like below, nice and tidy
-  // Short hands also available for answer query methods
-  teledart
-      .onMessage(keyword: 'Fight for freedom')
-      .listen((message) => message.reply('Stand with Hong Kong'));
 
   // Inline mode.
   teledart.onInlineQuery().listen((inlineQuery) => inlineQuery.answer([